SDR-304 feat: StoreDetail 페이지에서 Pictures 컴포넌트로 분리
