+++ fe/src/pages/StoreDetail/StoreDetail.tsx
@@ -1,41 +1,16 @@
-import { DEFAULT_IMG, STORE } from 'constants/dummyData';
-import { ICON, STORE_DETAIL } from 'constants/size';
-import TEXT from 'constants/text';
+import { STORE } from 'constants/dummyData';
+import { ICON } from 'constants/size';
 import CommonHeader from 'components/Common/Header/CommonHeader';
 import Icon from 'components/Common/Icon/Icon';
-import {
-  DimText,
-  InfoWrap,
-  MoreDim,
-  OtherPicture,
-  PictureWrap,
-  Row,
-  StoreInfo,
-  Text,
-  Title,
-  Wrap,
-} from './StoreDetail.styled';
+import { InfoWrap, Row, StoreInfo, Text, Title, Wrap } from './StoreDetail.styled';
 
 function StoreDetail() {
-  const { storeName, storeRating, storePictures, reviewCnt, address, phoneNumber } = STORE;
-  const [firstImg, ...otherImg] = getImagesOfStore({ pictures: storePictures, defaultImg: DEFAULT_IMG });
+  const { storeName, storeRating, reviewCnt, address, phoneNumber } = STORE;
 
   return (
     <>
       <CommonHeader />
       <Wrap>
-        <PictureWrap>
-          <div>
-            <img src={firstImg} alt={TEXT.ALT.FOOD} width={STORE_DETAIL.IMG.LARGE} height={STORE_DETAIL.IMG.LARGE} />
-          </div>
-          <OtherPicture>
-            {otherImg.map((picture) => (
-              <img src={picture} alt={TEXT.ALT.FOOD} width={STORE_DETAIL.IMG.SMALL} height={STORE_DETAIL.IMG.SMALL} />
-            ))}
-          </OtherPicture>
-          <MoreDim />
-          <DimText>더 보기</DimText>
-        </PictureWrap>
         <StoreInfo>
           <InfoWrap>
             <Title>{storeName}</Title>
@@ -57,11 +32,6 @@ function StoreDetail() {
       </Wrap>
     </>
   );
-
-  function getImagesOfStore({ pictures, defaultImg }) {
-    const ARR_LEN = 5;
-    return Array.from({ length: ARR_LEN }).map((_, i) => pictures[i] || defaultImg);
-  }
 }
 
 export default StoreDetail;
